feat: dashboard and workspace tweaks
